Address Michal's PR feedback
As Michal pointed out, SignatureContext no longer needs to implement
ISignatureTypeProvider so I removed it. I am not yet directly replacing
SignatureContext with ModuleTokenProvider as I believe their split
may be still useful for the envisioned merged multi-file R2R build.

In my mental view ModuleTokenProvider handles the synthetic token logic
and may ultimately become per input module whereas SignatureContext
semantically represents the [output] compilation module. I fully admit
my knowledge and design ideas in this area are still limited.

Based on Michal's PR feedback I have also switched TypesTableNode over
to use the same version-resilient hashcode calculation logic I implemented
while standing up the instance method table entrypoint R2R header table.

I have also moved the new test case authored by Michal to the
ReadyToRunInit test suite based on his PR suggestion; after all it's
just a small test case.

Thanks

Tomas
