+++ src/ILCompiler.ReadyToRun/src/Compiler/DependencyAnalysis/ReadyToRun/ModuleToken.cs
@@ -78,7 +78,7 @@ namespace ILCompiler.DependencyAnalysis.ReadyToRun
 
         public SignatureContext SignatureContext(ModuleTokenResolver resolver)
         {
-            return new SignatureContext(resolver, resolver.TypeSystemContext);
+            return new SignatureContext(resolver);
         }
 
         public MetadataReader MetadataReader => Module.PEReader.GetMetadataReader();
