+++ src/ILCompiler.ReadyToRun/src/Compiler/DependencyAnalysis/ReadyToRun/TypesTableNode.cs
@@ -46,7 +46,7 @@ namespace ILCompiler.DependencyAnalysis.ReadyToRun
                     rid = MetadataTokens.GetToken(ecmaType.Handle) & 0x00FFFFFF;
                     Debug.Assert(rid != 0);
 
-                    int hashCode = GetHashCode(ecmaType);
+                    int hashCode = ReadyToRunHashCode.TypeHashCode(ecmaType);
                     typesHashtable.Append(unchecked((uint)hashCode), section.Place(new UnsignedConstant((uint)rid << 1)));
                 }
                 else if (type.IsArray || type.IsMdArray)
@@ -68,26 +68,6 @@ namespace ILCompiler.DependencyAnalysis.ReadyToRun
                 relocs: null,
                 alignment: 8,
                 definedSymbols: new ISymbolDefinitionNode[] { this });
-        }
-
-        private int GetHashCode(EcmaType type)
-        {
-            int hashCode = ComputeNameHashCode(type);
-            while (type.ContainingType != null)
-            {
-                type = (EcmaType)type.ContainingType;
-                hashCode ^= ComputeNameHashCode(type);
-            }
-            return hashCode;
-        }
-        private int ComputeNameHashCode(EcmaType type)
-        {
-            int hashCode = TypeHashingAlgorithms.ComputeNameHashCode(type.Name);
-            if (!type.Namespace.Equals(""))
-            {
-                hashCode = TypeHashingAlgorithms.ComputeNameHashCode(type.Namespace) ^ hashCode;
-            }
-            return hashCode;
         }
 
         public override int ClassCode => -944318825;
