+++ src/ILCompiler.ReadyToRun/src/Compiler/ReadyToRunCodegenCompilationBuilder.cs
@@ -76,7 +76,7 @@ namespace ILCompiler
             var interopStubManager = new EmptyInteropStubManager(_compilationGroup, _context, new InteropStateManager(_context.GeneratedAssembly));
 
             ModuleTokenResolver moduleTokenResolver = new ModuleTokenResolver(_compilationGroup, _context);
-            SignatureContext signatureContext = new SignatureContext(moduleTokenResolver, _context);
+            SignatureContext signatureContext = new SignatureContext(moduleTokenResolver);
 
             ReadyToRunCodegenNodeFactory factory = new ReadyToRunCodegenNodeFactory(
                 _context,
