--- tests/src/Simple/ReadyToRunTokens/methodWhiteList.txt
@@ -1,5 +1,0 @@
-Program..cctor()
-Program+DisposeStruct..cctor()
-Program+DisposeClass.Dispose()
-Program+DisposeClass..cctor()
-ClassWithStatic..cctor()
