--- tests/src/Simple/ReadyToRunTokens/readytorun
@@ -1,1 +1,0 @@
-Execute this test in ready-to-run mode
